requirements.txt
-i https://pypi.org/simple
-e git+https://github.com/census-instrumentation/opencensus-python.git@85b462069440db7fe9455af4ce65752fe50acee5#egg=opencensus
cachetools==3.1.0
certifi==2019.3.9
chardet==3.0.4
click==7.0
flask==1.0.2
gevent==1.4.0 ; extra == 'gevent'
google-api-core[grpc]==1.8.2
google-auth==1.6.3
google-cloud-monitoring==0.31.1
googleapis-common-protos==1.5.9
greenlet==0.4.15 ; platform_python_implementation == 'CPython'
grpcio==1.19.0
gunicorn[gevent]==19.9.0
idna==2.8
itsdangerous==1.1.0
jinja2==2.10
markupsafe==1.1.1
protobuf==3.7.0
pyasn1-modules==0.2.4
pyasn1==0.4.5
pytz==2018.9
requests==2.21.0
rsa==4.0
six==1.12.0
urllib3==1.24.1
werkzeug==0.15.1
